Fixed compilation error on macOS
Non-constant expressions must be cast to the exact integer type when stored into a structure via a C-style initialization.

+++ source/sdl/i_sdlmusic.cpp
@@ -312,7 +312,8 @@ static void I_effectADLMIDIFloat(void *udata, Uint8 *stream, int len)
       return;
    }
 
-   const ADLMIDI_AudioFormat fmt = { ADLMIDI_SampleType_F32, ADLMIDISTEP, ADLMIDISTEP * audio_spec.channels };
+   const ADLMIDI_AudioFormat fmt = { ADLMIDI_SampleType_F32, ADLMIDISTEP,
+      static_cast<unsigned int>(ADLMIDISTEP * audio_spec.channels) };
 
    const int numsamples = (len * 2) / fmt.sampleOffset;
